requirements.txt
pip~=21.3.1
cryptography~=36.0.1
Pillow~=8.4.0
tornado~=6.1
Jinja2~=3.0.3
pytz~=2018.5
Django~=3.2.9
sqlparse~=0.2.4
psycopg2~=2.9.3
asgiref~=3.4.1
colorama~=0.4.4
numpy~=1.22.0
pipenv~=2021.11.23
ipython~=8.1.1
packaging~=21.3
distlib~=0.3.3
python-dateutil~=2.8.2
setuptools~=60.9.0
python-dotenv~=0.19.2
certifi~=2019.3.9
virtualenv~=20.10.0
six~=1.16.0
filelock~=3.4.0
platformdirs~=2.4.0
djangorestframework~=3.12.4
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
requests~=2.21.0
Pygments~=2.11.2